feat: add gateway api via cilium v1.18.0-pre.1 (#263)
Begin migration to gateway-api and moving away from nginx ingress since
the project is entering maintenance mode. To be able to set the
loadbalancerClass to Tailscale for the service cilium creates I need
access to `CiliumGatewayClassConfig` which is only in cilium 1.18 which
isn't released yet. The prerelease is manually pushed to the cluster and
the ks/hr for cilium are frozen.